requirements.txt
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
matplotlib
psutil
opencv-python
pillow
scikit-learn